requirement.txt
boto3
langchain
aws
streamlit
pillow
requests requests-aws4auth elasticsearch